+++ crates/swc/tests/tsc-references/nullishCoalescingOperator12.1.normal.js
@@ -1,7 +1,6 @@
 //// [nullishCoalescingOperator12.ts]
-var _obj;
 const obj = {
     arr: []
 };
 var _obj_arr;
-for (const i of (_obj_arr = (_obj = obj) === null || _obj === void 0 ? void 0 : _obj.arr) !== null && _obj_arr !== void 0 ? _obj_arr : []){}
+for (const i of (_obj_arr = obj === null || obj === void 0 ? void 0 : obj.arr) !== null && _obj_arr !== void 0 ? _obj_arr : []){}