+++ crates/swc/tests/tsc-references/nullishCoalescingOperator12.2.minified.js
@@ -1,2 +1,6 @@
 //// [nullishCoalescingOperator12.ts]
-for (const i of []);
+var _obj_arr;
+const obj = {
+    arr: []
+};
+for (const i of null !== (_obj_arr = null == obj ? void 0 : obj.arr) && void 0 !== _obj_arr ? _obj_arr : []);