+++ crates/swc_estree_compat/benches/babelify.rs
@@ -55,7 +55,7 @@ fn babelify_only(b: &mut Bencher) {
             module
                 .fold_with(&mut resolver(unresolved_mark, top_level_mark, true))
                 .fold_with(&mut typescript::strip(top_level_mark))
-                .fold_with(&mut es2020(Default::default()))
+                .fold_with(&mut es2020(Default::default(), unresolved_mark))
         });
 
         b.iter(|| {